upd_fdc: implement Terminal Count support
Add an API which allows signalling of the TC input to the FDC.  If TC
is asserted, store this in a flag and use it to disable sending further
result data for reads, or to fill in zeroes for writes.  Also disable
setting end-of-cylinder if TC is asserted.

Since the host usually asserts TC after reading/write the last data byte
to/from the data port, the start_{read,write}* functions are deferred
to an event, instead of being called directly from port I/O.  This gives
the host some time to assert TC.