if1_fdc: add 1kB RAM
When IF1 is paged in, map this RAM in the 8..16 kB region.  The mapping
is mirrored 8 times.  The 1st, 2nd, 5th and 6th kilobytes are read-only.
On the real HC, these would read the IF1 ROM; apart from a build ID
string (which we should eventually emulate), there is no useful data
there.