if1: disable microdrive I/O if if1_fdc is enabled
On the HC IF1, when both A3 and A4 are low, both the CTR and NET latches
are selected;  on read, this may be used for wire-anding the bits and
save one IN cycle.

Implement this behaviour just in case.  This also effectively disables
access to the original IF1 microdrive.