upd_fdc: fix writing of short sectors
If a write command is terminated before a sector is finished (e.g. due
to Terminal Count), the FDC should fill the remainder of the sector with
zeroes.

The code intends to do that but does a FDC_READ instead, such that the
fill doesn't happen.  Fix that by using FDC_WRITE.

Signed-off-by: Alex Badea <vamposdecampos@gmail.com>