cobra fdc: turn on motors
The CoBra FDC interface connects head load outputs to motor on pins; it
has no explicit control for motors.  For emulation, turn on the motors
all the time:  after a disk insert, and on reset.  The motor signal is
flipped on and off to make sure it triggers an update of the fdd ready
output.

Signed-off-by: Alex Badea <vamposdecampos@gmail.com>